pyproject.toml
[tool.poetry]
name = "ML_Blindness_Detection"
version = "0.1"
description = "ML Blind for IB."
repository = "https://github.com/Freddsle/ML_Blindness_Detection/"
authors = ["Yuliya Burankova, Leonid Zhozhikov, Dmitrii Podgalo, Artem Kosmin, Anton Zhelonkin"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
aiogram = "^2.20"
sklearn = "^0.0"
seaborn = "^0.11.2"
torch = "^1.11.0"
pandas = "^1.4.2"
matplotlib = "^3.5.2"
numpy = "^1.22.3"
Pillow = "^9.1.0"
torchvision = "^0.12.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"